weather-assistant-langchain/api-weather/requirements.txt
Flask==2.2.2
requests==2.28.1
gunicorn==20.1.0
pandas==1.1.5; python_version < '3.7'
pandas==1.3.5; python_version == '3.7'
pandas==2.0.1; python_version > '3.7'
pandas-gbq==0.19.2; python_version > '3.6'
# pandas-gbq==0.14.1 is the latest compatible version for Python 3.6
pandas-gbq==0.14.1; python_version < '3.7'
google-cloud-bigquery==3.12.0
#debugpy # Required for debugging.
#json
Werkzeug==2.2.2 